:recycle: refactoring smart contract
